+++ farmerMarket/storage/framework/views/c5d35b2770814346ac1cc528882337f0047cbfd8.php
@@ -1,22 +1,35 @@
 <?php $__env->startSection('content'); ?>
 <div class="container">
+
+
         <div class="row col-md-12">
             <div class="row col-md-8">
                 <div class="col-sm-6">
-                    <img src="<?php echo e(url('images/ads/' .$ads->id)); ?>" alt="Image Product" width="140" height="140" class="img-rounded"> </img> 
+                    <img src="<?php echo e(url('images/ads/' .$ads->id)); ?>" alt="Image Product" width="300" height="200" class="img"> </img> 
                 </div>
                 <div class="col-sm-6">
-                    <p><label class="control-label" for="owner">Owner: </label> <?php echo e($ads->owner_id); ?> </p>
+                    <p><label class="control-label" for="owner">Owner: </label> <?php echo e($ads->user->name); ?> </p>
 
                     <p><label class="control-label" for="name">Name: </label> <?php echo e($ads->name); ?></p>
+
+                    <p><label class="control-label" for="name">Description: </label> <?php echo e($ads->description); ?></p>
+
+                    <p><label class="control-label" for="name">Quantity: </label> <?php echo e($ads->quantity); ?></p>
+
+                    <p><label class="control-label" for="name">Trade Prefs: </label> <?php echo e($ads->trade_prefs); ?></p>
 
                     <p><label class="control-label" for="price">Open Price: </label> <?php echo e($ads->price_cents); ?>€ </p>
 
                     <p><label class="control-label" for="price">Minimal Price: </label> <?php echo e($ads->price_cents); ?>€ </p>
+
+                    <p><label class="control-label" for="name">Available Until: </label> <?php echo e($ads->available_until); ?></p>
                    
-                    <a href="/advertisement/bid/<?php echo e($ads->id); ?>"><button class="btn btn-primary">Bid</button></a> 
+
+                    <a href="/advertisement/bid/<?php echo e($ads->id); ?>"><button class="btn btn-primary">Bid</button></a>
+                    <?php if(Auth::user()->id == $ads->user->id): ?>
                     <a href="/advertisement/edit/<?php echo e($ads->id); ?>"><button class="btn btn-warning">Edit</button></a> 
-                    <a href="/advertisement/destroy/<?php echo e($ads->id); ?>"><button class="btn btn-danger">Del</button></a> 
+                    <a href="/advertisement/destroy/<?php echo e($ads->id); ?>" onclick="return confirm('Are you sure?')"><button class="btn btn-danger">Del</button></a> 
+                    <?php endif; ?>
 
                 </div>
             </div>
@@ -37,44 +50,17 @@
                         <i class="fa fa-commenting"></i> New comment
                 </button>
             </div>
+            <div class="row col-md-12">
+                <?php echo $__env->make('comments.comments_view', array_except(get_defined_vars(), array('__data', '__path')))->render(); ?>
+            </div>
         </div>
 
-            <div class="row col-md-12">
-                    <?php if(count($comments) > 0): ?>
-                        <?php foreach($comments as $comment): ?>
-                        <div class="row">
-                            <div class="col-md-2">
-                                <label class="control-label" for="user">User: </label>
-                                <?php echo e($comment->author); ?>
 
-                            </div>
-                            <div class="col-md-4">
-                                <label class="control-label" for="date">Date: </label>
-                                <?php echo e($comment->created_at); ?>
-
-                            </div>
-                        </div>
-                         <div class="row">
-                            <div class="col-md-6">
-                                <label class="control-label" for="user">Mensage: </label>
-                                <?php echo e($comment->comment); ?>
-
-                            </div>
-                            <div class="col-md-2">
-                                <button data-toggle="modal" data-target="#replayComment" class="btn btn-warning">
-                                        <i class="fa fa-mail-reply"></i> Replay
-                                </button>
-                            </div>
-                                <br /><hr>
-                        </div>
-                        <?php endforeach; ?>
-                    <?php else: ?>
-                        No comments.. Be the First :)
-                    <?php endif; ?>
-
-                
-            </div>
 </div>
+
+
+<?php $__env->stopSection(); ?>
+
 
 <div id="newComment" class="modal fade" role="dialog">
     <div class="modal-dialog">
@@ -104,13 +90,24 @@
             <h4 class="modal-title">Replay Comment</h4>
           </div>
             <div class="modal-body">
-            <form class="form-horizontal" role="form" method="POST" action="<?php echo e(url('/comment/reply')); ?>">
+            <form class="form-horizontal" role="form" method="POST" action="<?php echo e(url('/comment/new')); ?>">
+            <input type="hidden" class="form-control" id="parent_id" name="parent_id" value="" />
                 <?php echo $__env->make('comments.comment_form', array_except(get_defined_vars(), array('__data', '__path')))->render(); ?>
 
             </div>
         </div>
     </div>
 </div>
-<?php $__env->stopSection(); ?>
+
+<script type="text/javascript">
+$('#replayComment').on('show', function(e) {
+    alert("sdfsdf");
+    //get data-id attribute of the clicked element
+    var parent_id = $(e.relatedTarget).data('id');
+
+    //populate the textbox
+    $(e.currentTarget).find('input[name="parent_id"]').val(parent_id);
+});
+</script>
 
 <?php echo $__env->make('layouts.app', array_except(get_defined_vars(), array('__data', '__path')))->render(); ?>
